Store read lock holders under separate ETS keys
The old way of storing all read lock holders in a list meant releasing
a lock was an O(n) operation. From now on the locks are stored as
separate objects in the ETS (hash)table, so lookup and delete
operations will become O(1). This has a significant impact when there
are many locks in the system.